Add or remove metadata

In iRODS, metadata is stored as attribute-value-units triples (AVUs), consisting of an attribute name, an attribute value and an optional unit. This function allows to chain several operations ('add' or 'remove') linked to specific AVUs.

Usage

imeta(
  logical_path,
  entity_type = c("data_object", "collection", "user"),
  operations = list(),
  verbose = FALSE
)

Arguments

logical_path

Path to the data object or collection (or name of the user).

entity_type

Type of item to add metadata to or remove it from. Options are 'data_object', 'collection' and 'user'.

operations

List of named lists of vectors representing operations. The valid components of each of these lists or vectors are:

operation, with values 'add' or 'remove', depending on whether the AVU should be added to or removed from the metadata of the item.

attribute, with the name of the AVU.

value, with the value of the AVU.

units, with the unit of the AVU.

verbose

Whether information should be printed about the HTTP request and response.

Value

Invisibly, the HTTP response.

Examples

if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo, "foo.rds")

# check if file is stored
ils()

# add some metadata
imeta(
 "foo.rds",
 "data_object",
 operations = list(
    operation = "add",
    attribute = "foo",
    value = "bar",
    units = "baz"
    )
)

imeta(
 "foo.rds",
 "data_object",
 operations = list(
   list(operation = "add", attribute = "foo2", value = "bar2"),
   list(operation = "add", attribute = "foo3", value = "bar3")
 )
)

# check if file is stored with associated metadata
ils(metadata = TRUE)
}