Save files and objects in iRODS

Store an object or file into iRODS. iput() should be used to transfer a file from the local storage to iRODS; isaveRDS() saves an R object from the current environment in iRODS in RDS format (see saveRDS()).

Usage

iput(
  local_path,
  logical_path = basename(local_path),
  offset = 0,
  count = 2000L,
  truncate = "true",
  verbose = FALSE,
  overwrite = FALSE
)

isaveRDS(
  x,
  logical_path,
  offset = 0,
  count = 2000L,
  truncate = "true",
  verbose = FALSE,
  overwrite = FALSE
)

Arguments

local_path

Local path of file to be sent to iRODS.

logical_path

Destination path in iRODS. By default, the basename of local_path is used and the file is stored in the current working directory (see ipwd()).

offset

Offset in bytes into the data object. Defaults to 0.

count

Maximum number of bytes to write. Defaults to 2000.

truncate

Whether to truncate the object when opening it. Defaults to TRUE.

verbose

Whether to print information about the HTTP request and response. Defaults to FALSE.

overwrite

Whether the file in iRODS should be overwritten if it exists. Defaults to FALSE.

x

R object to save in iRODS.

Value

(Invisibly) the HTTP response.

Examples

if (interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home", TRUE)

# authenticate
iauth()

# save the iris dataset as csv and send the file to iRODS
library(readr)
write_csv(iris, "iris.csv")
iput("iris.csv", overwrite = TRUE)

# save with a different name
iput("iris.csv", "irids_in_irods.csv", overwrite = TRUE)
ils()

# send an R object to iRODS in RDS format
isaveRDS(iris, "irids_in_rds.rds", overwrite = TRUE)
}